+++ Cargo.toml
@@ -23,6 +23,7 @@ toml = "0.9.0"
 procfs = "0.18.0"
 uom = "0.37.0"
 udev = { version = "0.9.3", features = ["hwdb"]}
+libc = "0.2.181"
 
 [profile.dev.package.backtrace]
 opt-level = 3
